Added Expo Demo, small updated to Sample and Screenshots
This commit adds an Expo snack link so that it is easier for first timers to be able to see the code as well as see what it looks like. Expo is a great tool for prototyping and this demo will definitely help newcomers to React Native use this plugin with ease. Next steps would be to add a Medium article where a more detailed explanantion can be provided and working with the plugin can be super easy.