Improved sizing for ConfirmDialogs that contain ScrollViews
ConfirmDialogs with a ScrollView containing a large amount of content
are often displayed with a height larger than the available screen
space, so that the bottom of the dialog is cut off.

This commit fixes that issue.

The idea to use flex: -1 comes from
medium.com/@peterpme/taming-react-natives-scrollview-with-flex-144e6ff76c08